log error for NaN SVObs values
if the "--allow-non-numeric-svobs" flag is false and a SVObs "value"
property value is non-numeric, log an error.

note that while the CLI flag defaults to false, calling McfChecker
without arguments defaults to allowing NaN SVObs. This is so that
minimal amount of internal tests fail until we want to modify that
behavior.