only test date patterns that are potential matches based on their length
We use exceptions to determine invalid matches of date patterns in
StringUtil. This has a high CPU cost as every exception requires filling
the stack trace. This change reduces the potential number of failures by
filtering out changes that we know won't work with a length-based
heuristic.

In testing, I used open-covid-19 data available at [1]. I ran the tool
on the first 10k rows of the open_covid_19_data.csv file, with the other
.mcf and .tmcf files.

Over 3 runs before and after this commit, I saw the time average time
taken by the tool from 53.2 to 48.3 seconds. This is a speed-up of 9%.

Note that the chosen dataset has dates in the format YYYY-MM-dd, and it
does not have any invalid dates. The expected speed-up from this change
is higher when there are many invalid dates, when the date formats are
longer (since the previous approach tries roughly increasing pattern
lenghts). The expected speed-up is less if the date format is shorter,
but never negative (ie, this change will not lead to more checks being
done than the previous approach).

[1] https://github.com/datacommonsorg/data/tree/master/scripts/google_research/open-covid-19