clean up implementation to use a single SoT
A single variable of type Map<String, List<Integer>> is used instead of
adding more logic to construct an instant map. The speed gains from this
PR are that we are throwing less exceptions (which this commit does not
break); the trivial for-loop computation is not relevant to the speed
here.