Filter out impossible date pattern checks for performance gains (#173)
* only test date patterns that are potential matches based on their length

We use exceptions to determine invalid matches of date patterns in
StringUtil. This has a high CPU cost as every exception requires filling
the stack trace. This change reduces the potential number of failures by
filtering out changes that we know won't work with a length-based
heuristic.

In testing, I used open-covid-19 data available at [1]. I ran the tool
on the first 10k rows of the open_covid_19_data.csv file, with the other
.mcf and .tmcf files.

Over 3 runs before and after this commit, I saw the time average time
taken by the tool decrease from 45.144s to 33.565s. This is an estimated speed-up of ~26%.

[1] https://github.com/datacommonsorg/data/tree/master/scripts/google_research/open-covid-19