Add support for slashes in entity type names (#393)
Also fix attribute name (again)

This fixes a bug that caused new entity type names with slashed prefixes
to crash the data management job. It also changes the behavior so that
slash-prefixed stat vars and entity types are not considered "missing"
DCIDs and doesn't cause the importer to auto-create DCIDs.