Fix parent versions in pom.xml (#421)
Follow up to
https://github.com/datacommonsorg/import/commit/4c2964f0c1961c084d8ec1ca0952ce08dd15cede

It looks like parent versions need to be explicitly set to compile the
ingestion module to run the ingestion pipeline. (Apparently revision is
a special property that allows this, but custom properties do not)

To keep the stable modules (0.1-alpha.1) separate from the development
modules (0.1-SNAPSHOT, which is currently set for revision), I have to
set the parent versions directly.

Tested by running ingestion pipeline